requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.7.4
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
